The `glob` lookup for the file `CertificateAuthorityCertificate.pem` causes a fairly significant performance slowdown when issuing any request, and it worsens the more files are present in the project.  By replacing this with a simple, synchronous exists check, we can realize fairly significant performance improvements.  

On a current project, I was seeing ~3-second delays for every interaction through the library, while the same commands via curl completed in tens of milliseconds.  This simple change cut cycle time for my tests from ~9 seconds per test to ~200ms.  

But since you can't easily reproduce my results, we can use the test suite for this project as proof.  This is the result of running the grunt suite before the improvement:
```
brian:~/work_repos/mockserver-client-node$ time node_modules/grunt/bin/grunt
Running "exec:stop_existing_mockservers" (exec) task
done
...
Chrome 78.0.3904 (Linux 0.0.0): Executed 34 of 34 SUCCESS (4.607 secs / 4.622 secs)
TOTAL: 34 SUCCESS


Running "stop_mockserver" task

Done.

real    6m25.084s
user    5m34.983s
sys     0m28.594s
```

And here are the same tests after this change:
```
brian:~/work_repos/mockserver-client-node$ time node_modules/grunt/bin/grunt
Running "exec:stop_existing_mockservers" (exec) task
done
...
Chrome 78.0.3904 (Linux 0.0.0): Executed 34 of 34 SUCCESS (4.841 secs / 4.852 secs)
TOTAL: 34 SUCCESS


Running "stop_mockserver" task

Done.

real    0m37.302s
user    0m5.040s
sys     0m0.848s
```