@jamesdbloom 2019-11-16
This looks like a good improvement.  I'll be doing a new release in the next few days which will include this change.